requirements.txt

# Core AI Libraries (Multiple model support)
google-generativeai>=0.4.0
openai>=1.0.0
ollama>=0.1.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-community>=0.1.0
requests>=2.28.0

# Document processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
pypandoc>=1.11

# Text processing and embeddings
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.20.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
